DiverMeal-main/pythonBackend/requirements.txt
Flask==1.1.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
keras==2.7.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.12.2
skimage==0.0
tensorflow==2.7.0
